route_d/requirements.txt
livekit-agents[groq,silero]>=1.5.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
